+++ CHANGES.md
@@ -12,7 +12,7 @@ This is a history of changes to [Remix](https://remix.run).
   - `createCookieSessionStorage`
   - `createFileSessionStorage`
   - `createMemorySessionStorage`
-- Added `createSessionStorage` for easily implementing the `SessionStorage` interface
+- Added `createSessionStorage` for easily implementing the `SessionStorage` interface using `SessionIdStorageStrategy`
   using any database
 - Added `Cookie` interface and `createCookie` API for handling cookies, including
   support for signed cookies